% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloader.R
\name{getExprSet}
\alias{getExprSet}
\title{download exprSet data}
\usage{
getExprSet(file = "", outdir = ".", url = "119.29.67.25",
  user = NULL, password = NULL, dir = "pub")
}
\arguments{
\item{file}{one of pancancer_mRNA_exprSet.Rdata
panCCLE_mRNA_exprSet.Rdata  panGTEX_mRNA_exprSet.Rdata}

\item{url}{ftp url}

\item{user}{ftp user}

\item{password}{ftp password}

\item{dir}{ftp directory}
}
\description{
download exprSet data
}
\examples{
temp_dir <- tempdir()
getExprSet(file = "welcome.txt", dir = temp_dir)

\dontrun{
getExprSet(file = "pancancer_mRNA_exprSet.Rdata")
}

}
\author{
Zhougeng Xu
}
