% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pancancer_small}
\alias{pancancer_small}
\title{TCGA preprocessed data for correlationship analysis}
\format{data.frame object}
\usage{
data(pancancer_small)
}
\description{
Data are downloaded from \href{http://xena.ucsc.edu/}{XENA}
}
\examples{
data(pancancer_small)
}
\keyword{datasets}
