% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{panCCLE_small}
\alias{panCCLE_small}
\title{CCLE preprocessed data for correlationship analysis}
\format{data.frame object}
\usage{
data(panCCLE_small)
}
\description{
Data are downloaded from \href{https://portals.broadinstitute.org/ccle}{CCLE}
}
\examples{
data(panCCLE_small)
}
\keyword{datasets}
