% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pancor.R
\name{pancor}
\alias{pancor}
\title{caculate the correlatship ratio in pancancer}
\usage{
pancor(data, gene1, gene2, cor.test.method = "pearson")
}
\arguments{
\item{data}{data.frame}

\item{gene1}{charactor of gene name, e.g. "METTL3"}

\item{gene2}{charactor of gene name, e.g. "SETD2"}

\item{cor.test.method}{cor.test method, one of
"pearson", "kendall", "spearman", default is "pearson"}
}
\description{
caculate the correlatship ratio in pancancer
}
\examples{
data(pancancer_small)
pancor(pancancer_small, "METTL3", "SETD2")

}
\author{
Shipeng Guo, Zhougeng Xu
}
