% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{panGTEX_small}
\alias{panGTEX_small}
\title{GTEx preprocessed data for correlationship analysis}
\format{data.frame object}
\usage{
data(panGTEX_small)
}
\description{
Data are downloaded from \href{https://gtexportal.org/home/}
}
\examples{
data(panGTEX_small)
}
\keyword{datasets}
