% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloader.R
\name{getExprSet}
\alias{getExprSet}
\title{download exprSet data}
\usage{
getExprSet(file = "", outdir = ".", dir = "pub", url = NULL,
  user = NULL, password = NULL)
}
\arguments{
\item{file}{one of pancancer_mRNA_exprSet.Rdata
panCCLE_mRNA_exprSet.Rdata  panGTEX_mRNA_exprSet.Rdata}

\item{outdir}{data store directory}

\item{dir}{ftp directory}

\item{url}{ftp url}

\item{user}{ftp user}

\item{password}{ftp password}
}
\value{
a .Rdata file in working directory
}
\description{
download exprSet data
}
\examples{
temp_dir <- tempdir()
getExprSet(file = "test.txt", outdir = temp_dir)

\donttest{
# According to CRAN, function cannot modify user's home filespace,
# so when run the example, the data will save at the tempdir.
# You can remove the `outdir = tempdir()` to let the .Rdata save to
# your current directory
getExprSet(file = "pancancer_mRNA_exprSet.Rdata", outdir = tempdir())
}

}
\author{
Zhougeng Xu
}
