% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{ggcorplot}
\alias{ggcorplot}
\title{ggcorplot}
\usage{
ggcorplot(data, gene1, gene2, type = "ALL", cor.test.method = "person")
}
\arguments{
\item{data}{data.frame}

\item{gene1}{charactor of gene name, e.g. "METTL3"}

\item{gene2}{charactor of gene name, e.g. "SETD2"}

\item{type}{a kind of cancer, like BRCA, default is ALL}

\item{cor.test.method}{cor.test method, one of
"pearson", "kendall", "spearman", default is "pearsson"}
}
\description{
ggcorplot
}
\examples{
data()
data(pancancer_small)
ggcorplot(pancancer_small, "METTL3","SETD2")
ggcorplot(pancancer_small, "METTL3","SETD2", "BRCA")

}
\author{
Shipeng Guo, Zhougeng Xu
}
